Make all times same format and other changes
- Times would often be either %d-%b-%Y or '%b-%d-%Y. Now they are all '%d-%b-%Y
- Dates to sort logs by now include break log dates
- Fix sorting in on call logs
- Move driver log functions to a driver log controller
